add builder for retry logic and static entry point to ease usage
